backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "openai>=2.32.0",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "resend>=2.29.0",
    "supabase>=2.29.0",
    "uvicorn[standard]>=0.46.0",
]
